requirements.txt
absl-py==2.3.0
aiofiles==24.1.0
annotated-types==0.7.0
argcomplete==3.5.1
asttokens==3.0.0
attrs==24.2.0
Cerberus==1.3.5
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
chex==0.1.89
clarabel==0.11.1
click==8.2.1
cloudpickle==3.1.1
colour==0.1.5
contourpy==1.3.1
control==0.10.1
coverage==7.6.8
cvxpy==1.6.6
cycler==0.12.1
decorator==5.2.1
dm-tree==0.1.9
docker-pycreds==0.4.0
einops==0.8.1
empy==3.3.4
equinox==0.13.1
etils==1.12.2
exceptiongroup==1.3.0
executing==2.2.0
flax==0.10.7
fonttools==4.55.0
fsspec==2025.5.1
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
humanize==4.12.3
idna==3.10
importlib_resources==6.4.5
ipdb==0.13.13
ipython==8.36.0
jax==0.6.2
jax-cuda12-pjrt==0.6.2
jax-cuda12-plugin==0.6.2
jaxlib==0.6.2
jaxopt==0.8.5
jaxproxqp @ git+https://github.com/oswinso/jaxproxqp.git@8b6d76028ef19642b9684a33584624295ea69128
jaxtyping==0.3.2
jedi==0.19.2
Jinja2==3.1.4
joblib==1.5.1
jraph==0.0.6.dev0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kconfiglib==14.1.0
kiwisolver==1.4.7
loguru==0.7.3
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
msgpack==1.1.0
nest-asyncio==1.6.0
numpy==2.2.6
nunavut==2.3.1
nvidia-cublas-cu12==12.9.0.13
nvidia-cuda-cupti-cu12==12.9.19
nvidia-cuda-nvcc-cu12==12.9.41
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.37
nvidia-cudnn-cu12==9.10.1.4
nvidia-cufft-cu12==11.4.0.6
nvidia-cusolver-cu12==11.7.4.40
nvidia-cusparse-cu12==12.5.9.5
nvidia-nccl-cu12==2.26.5
nvidia-nvjitlink-cu12==12.9.41
nvidia-nvshmem-cu12==3.2.5
opencv-python==4.11.0.86
opt_einsum==3.4.0
optax==0.2.4
orbax-checkpoint==0.11.25
osqp==1.0.4
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
pkgconfig==1.5.5
platformdirs==4.3.8
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydsdl==1.22.0
Pygments==2.19.1
pymavlink==2.4.42
pyparsing==3.2.3
pyros-genmsg==0.5.8
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.2
pyulog==1.1.0
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==14.0.0
rpds-py==0.21.0
scs==3.2.7.post2
seaborn==0.13.2
sentry-sdk==2.29.1
setproctitle==1.3.6
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.13.3
tensorflow-probability==0.25.0
tensorstore==0.1.75
toml==0.10.2
tomli==2.2.1
toolz==1.0.0
tqdm==4.67.1
traitlets==5.14.3
treescope==0.1.9
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2024.2
urllib3==2.4.0
wadler_lindig==0.1.6
wandb==0.19.11
wcwidth==0.2.13
wrapt==1.17.2
zipp==3.22.0
